packages/solver/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sinanz"
version = "0.1.0"
description = "Standalone Python solver package for click and slider captchas."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "numpy",
  "onnxruntime",
  "pillow",
]

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = [
  "sinanz",
  "sinanz_errors",
  "sinanz_group1_runtime",
  "sinanz_group1_service",
  "sinanz_group2_runtime",
  "sinanz_group2_service",
  "sinanz_image_io",
  "sinanz_resources",
  "sinanz_types",
]
include-package-data = true

[tool.setuptools.data-files]
"resources" = ["resources/manifest.json"]
"resources/metadata" = [
  "resources/metadata/README.md",
  "resources/metadata/*.json",
]
"resources/models" = [
  "resources/models/README.md",
  "resources/models/*.onnx",
  "resources/models/*.onnx.data",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.12"
files = ["src", "tests"]
strict = true
warn_unused_configs = true

[tool.sinanz]
public_api = ["sn_match_slider", "sn_match_targets"]
runtime_target = "python-onnxruntime"

[dependency-groups]
dev = [
  "mypy>=1.11",
  "pytest>=8.3",
  "ruff>=0.6",
]
